+++ CHANGELOG.md
@@ -10,3 +10,13 @@ All notable changes to this project will be documented in this file.
 - Changed: package.json
 - Changed: src/
 - Changed: tsconfig.lib.json
+
+## [2.0.0] - 2025-04-29
+
+- Modified: package.json
+- Modified: src/auth.module.ts
+- Modified: src/guards/redis.guard.ts
+- Modified: src/serializers/session.serializer.ts
+- Modified: src/web2/src/auth/web2.service.ts
+- Modified: src/web3/src/auth/web3.controller.ts
+- Modified: src/web3/src/guards/web3.guard.ts